analyzer/sheratan-reconciler-sheratan-conform/requirements.txt
# Core dependencies
click>=8.1.0
pyyaml>=6.0
colorama>=0.4.6
rich>=13.0.0

# File and code analysis
gitpython>=3.1.0
pathspec>=0.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
